refactor: Change the integrity package structure
- pkg structure based on functionality
