+++ app/src/main/java/uk/gov/android/authentication/integrity/ClientAttestationManager.kt
@@ -1,9 +1,9 @@
 package uk.gov.android.authentication.integrity
 
-import uk.gov.android.authentication.integrity.model.AttestationResponse
-import uk.gov.android.authentication.integrity.model.SignedResponse
+import uk.gov.android.authentication.integrity.appcheck.model.AttestationResponse
+import uk.gov.android.authentication.integrity.pop.SignedResponse
 
 interface ClientAttestationManager {
     suspend fun getAttestation(): AttestationResponse
-    suspend fun generatePoP(iss: String, aud: String): SignedResponse
+    fun generatePoP(iss: String, aud: String): SignedResponse
 }
