+++ app/src/main/java/uk/gov/android/authentication/integrity/FirebaseClientAttestationManager.kt
@@ -1,16 +1,16 @@
 package uk.gov.android.authentication.integrity
 
 import android.util.Log
-import uk.gov.android.authentication.integrity.appcheck.AppChecker
+import uk.gov.android.authentication.integrity.appcheck.usecase.AppChecker
 import uk.gov.android.authentication.integrity.keymanager.ECKeyManager
 import uk.gov.android.authentication.integrity.keymanager.KeyStoreManager
-import uk.gov.android.authentication.integrity.model.AppCheckToken
+import uk.gov.android.authentication.integrity.appcheck.model.AppCheckToken
+import uk.gov.android.authentication.integrity.appcheck.model.AttestationResponse
+import uk.gov.android.authentication.integrity.pop.ProofOfPossessionGenerator
+import uk.gov.android.authentication.integrity.pop.SignedResponse
+import uk.gov.android.authentication.integrity.appcheck.usecase.AttestationCaller
 import uk.gov.android.authentication.integrity.model.AppIntegrityConfiguration
-import uk.gov.android.authentication.integrity.model.AttestationResponse
-import uk.gov.android.authentication.integrity.model.ProofOfPossessionGenerator
-import uk.gov.android.authentication.integrity.model.SignedResponse
-import uk.gov.android.authentication.integrity.usecase.AttestationCaller
-import uk.gov.android.authentication.integrity.usecase.JWK
+import uk.gov.android.authentication.integrity.appcheck.usecase.JWK
 import java.security.SignatureException
 import kotlin.io.encoding.ExperimentalEncodingApi
 
@@ -41,7 +41,7 @@ class FirebaseClientAttestationManager(
         }
     }
 
-    override suspend fun generatePoP(iss: String, aud: String): SignedResponse {
+    override fun generatePoP(iss: String, aud: String): SignedResponse {
         // Create Proof of Possession
         val pop = ProofOfPossessionGenerator.createBase64PoP(iss, aud)
         // Convert into ByteArray
