--- app/src/main/java/uk/gov/android/authentication/integrity/appcheck/AppChecker.kt
@@ -1,7 +1,0 @@
-package uk.gov.android.authentication.integrity.appcheck
-
-import uk.gov.android.authentication.integrity.model.AppCheckToken
-
-fun interface AppChecker {
-    suspend fun getAppCheckToken(): Result<AppCheckToken>
-}
