+++ app/src/main/java/uk/gov/android/authentication/integrity/appcheck/model/AttestationResponse.kt
@@ -1,4 +1,4 @@
-package uk.gov.android.authentication.integrity.model
+package uk.gov.android.authentication.integrity.appcheck.model
 
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.Serializable
