+++ app/src/main/java/uk/gov/android/authentication/integrity/appcheck/usecase/AppChecker.kt
@@ -1,0 +1,7 @@
+package uk.gov.android.authentication.integrity.appcheck.usecase
+
+import uk.gov.android.authentication.integrity.appcheck.model.AppCheckToken
+
+fun interface AppChecker {
+    suspend fun getAppCheckToken(): Result<AppCheckToken>
+}
