+++ app/src/main/java/uk/gov/android/authentication/integrity/model/AppIntegrityConfiguration.kt
@@ -1,8 +1,8 @@
 package uk.gov.android.authentication.integrity.model
 
-import uk.gov.android.authentication.integrity.appcheck.AppChecker
+import uk.gov.android.authentication.integrity.appcheck.usecase.AppChecker
 import uk.gov.android.authentication.integrity.keymanager.KeyStoreManager
-import uk.gov.android.authentication.integrity.usecase.AttestationCaller
+import uk.gov.android.authentication.integrity.appcheck.usecase.AttestationCaller
 
 data class AppIntegrityConfiguration(
     val attestationCaller: AttestationCaller,
