+++ app/src/test/java/uk/gov/android/authentication/integrity/appcheck/usecase/JWKTest.kt
@@ -1,5 +1,6 @@
-package uk.gov.android.authentication.integrity.usecase
+package uk.gov.android.authentication.integrity.appcheck.usecase
 
+import uk.gov.android.authentication.integrity.appcheck.usecase.JWK
 import kotlin.test.Test
 import kotlin.test.assertEquals
