+++ app/src/test/java/uk/gov/android/authentication/integrity/pop/ProofOfPossessionGeneratorTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.android.authentication.integrity.model
+package uk.gov.android.authentication.integrity.pop
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -6,7 +6,7 @@ import kotlin.test.assertEquals
 class ProofOfPossessionGeneratorTest {
     @Test
     fun `create Base64 Proof of Possession`() {
-        val expectedResult = ClassLoader.getSystemResource("./bodyPoPBase64.txt").readText()
+        val expectedResult = ClassLoader.getSystemResource("bodyPoPBase64.txt").readText()
         val result = ProofOfPossessionGenerator.createBase64PoP(
             "iss",
             "aud",
